// apps/website/src/types/documentation.ts
type DocumentationClassMethodParameterMeta = DocumentationClassMeta;
type DocumentationClassMethodPropertyMeta = DocumentationClassMeta;
type DocumentationClassEventMeta = DocumentationClassMeta;
type DocumentationExternalMeta = DocumentationClassMeta;
type DocumentationTypeDefinitionMeta = DocumentationClassMeta;
type DocumentationTypeDefinitionParameterMeta = DocumentationClassMeta;

interface DocumentationParameter {
	name: string;
	description?: string;
	default: string;
	abstract?: boolean;
	deprecated?: boolean | string;
	access?: string;
	readonly?: boolean;
	optional?: boolean;
	variable?: boolean;
	nullable?: boolean;
	scope?: string;
	see?: string[];
	type: string[][][];
}
type DocumentationClassConstructorParameter = DocumentationParameter;
type DocumentationClassEventParameter = DocumentationParameter;

interface DocumentationClassMethodParameter extends DocumentationParameter {
	meta: DocumentationClassMethodParameterMeta;
}

interface DocumentationTypeDefinitionParameter extends DocumentationParameter {
	meta: DocumentationTypeDefinitionParameterMeta;
}
// ...
export type ParameterUnion =
	| DocumentationClassConstructorParameter
	| DocumentationClassMethodParameter
	| DocumentationClassEventParameter
	| DocumentationTypeDefinitionParameter;